twitter_sentiment_analysis/requirements.txt
appdirs==1.4.4
black==21.4b2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
flake8==3.9.1
fpdf==1.7.2
idna==2.10
isort==5.8.0
joblib==1.0.1
kaleido==0.2.1
kiwisolver==1.3.1
matplotlib==3.4.1
mccabe==0.6.1
mypy-extensions==0.4.3
nltk==3.6.2
numpy==1.20.2
oauthlib==3.1.0
pandas==1.2.4
pathspec==0.8.1
Pillow==8.2.0
plotly==4.14.3
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
textblob==0.15.3
toml==0.10.2
tqdm==4.60.0
tweepy==3.10.0
urllib3==1.26.4
wordcloud==1.8.1
